pyproject.toml
[tool.poetry]
name = "aircmd"
version = "0.1.2"
description = "Airbyte CLI"
authors = ["cpdeethree <conor@airbyte.io>"]


[tool.poetry.dependencies]
python = "^3.11"
dagger-io = "^0.6.0"
structlog = "^23.1.0"
rich = "^13.3.5"
pydantic = {version = "^2.0b2", allow-prereleases = true}
pydantic-settings = { git = "https://github.com/pydantic/pydantic-settings.git", branch = "main" }
requests = "^2.26.0"
asyncclick = "^8.1.3.4"

[tool.mypy]
strict = "true"

[tool.ruff]
# ignore line length violations for now
ignore = ["E501"]

[tool.poetry.scripts]
aircmd = "aircmd.main:main"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.262"
mypy = "^1.2.0"

#[tool.poetry.dev.dependencies]
# Example plugin goes here. We run tests against it.
# This also doubles as the official plugin registry
types-requests = "^2.31.0.1"
[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

